Refactor damage handling to support negative damage values and update related types

+++ src/core/game/helper/dmg.rs
@@ -2,13 +2,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct Dmg {
-    pub amount: u64,
+    pub amount: i32,
     pub attacker_id: u64,
     pub target_id: u64,
 }
 
 impl Dmg {
-    pub fn new(attacker_id: u64, target_id: u64, dmg: u64) -> Self {
+    pub fn new(attacker_id: u64, target_id: u64, dmg: i32) -> Self {
         Self {
             attacker_id,
             target_id,
